VMS-55323: Simplify user settings saving
Requirement of the fresh session token was removed.

`UserSettingsEx` structure was removed as well because settings stored there were already moved to
a separate property (not required in the User API).

Helper functions were moved out of the UserResource class to the usage places, so duplicating
`UserSettings` class was removed.