MOBILE-2136: iOS: Restore manual provisioning profile copying for non-developer builds
Xcode can embed the profile into the application archive, but we do not use it for creating and
exporting the archives. We prepare IPAs manually and upload them via Transporter. Thus we have to
embed the profile manually.

This does not affect development builds which are deployed directly to a device by Xcode.