VMS-55838: Remove version mismatch warning
This warning appeared a long time ago when we had a dedicated component for storing the system
database. That time we did not have automated system updates and did not have strict compatibility
requirements between the components, so users could get in situations when the system components
had different versions. In current installations cases when Mediaservers are connected to each
other but have different versions are very rare and not really harmful, so this warning message is
no longer valid. It mostly complains on a Client which has newer version than the connected
Mediaserver, but such situations are normal and will happen regularly once we decide to make a
client-only update.