VMS-55967: Fix double cloud authorization after system settings changes
Closes VMS-55967

When performing an API request that requires recent password input, the client prompts the user to complete OAuth authorization. After obtaining a fresh token, it was stored only within the `ServerConnection`, without updating the credentials in `RemoteSession or `CloudStatusWatcher`. This led to a situation where, upon making an API request to reduce the duration of a cloud session, the user was logged out of the system, despite having just entered their login and password.

Now the `refresh_token` issued after confirming the API request is passed to the `CloudStatusWatcher`

A critical point: CDB allows issuing both `refresh_token` and `access_token` at once, but both must either have a scope or lack it entirely. For `CloudStatusWatcher`, tokens without a scope are required, whereas for `RemoteSession` in `ServerConnection`, an `access_token` with a scope is needed. Since I am unaware of any cases where a `refresh_token` with a scope is necessary, I first issue a `refresh_token` and `access_token` without a scope. If required, I then issue an additional `access_token` with a scope.