VMS-56362: Revert "Add validation to lookup list creation with REST API" and its related fixes
The validation logic for lookup list creation is being removed because it was a direct copy-paste of client-side code. However, this implementation does not account for derived ObjectTypes, which can include additional data(new attributes, new possible values) to the base ObjectType (as handled correctly in the desktop client). As a result, the server-side validation behaves incorrectly.

A common solution that works consistently across both client and server needs to be implemented to address this issue properly.