VMS-55104, VMS-55597: Cleanup old Rules Engine from code, implement compatibility layer for Event Log v3
Old rules engine was partially removed from code. The complete cleanup will be done in the
following commits separately.

Implemented compatibility layer and the corresponding structures for the old event log
handlers (including `/api/getEvents`, `/ec2/getEvents` and `/rest/v3/events/log`).

Enabled audit trail logging for the new rules transactions. Format is similar to the old
engine, but locale-independent. Conversion to the translatable human-readable representation
is done on the client side in the dialog.