VMS-56971: Do not request Conan updates unless it is required to update a recipe revision
We can greatly simplify Conan execution and remove all the optimizations for improperly revisioned
packages because we no longer have such packages. Each package either has a unique version which
should be changed to pull an update or it has a unique recipe revision which also needs to be
changed to pull an update.