VMS-56371: Fix server assertion failure in layout_item_watcher.cpp
Description of the problem I fixed here.

We call `LayoutItemWatcher::addWatchedLayout` from a handler of layout's `QnResource::parentIdChanged` signal (when the layout becomes shared). 

Then it fetches layout items and connects to `QnLayoutResource::itemAdded`/`itemRemoved`/`itemChanged`.

However there can be a situation when both parentId and layout items are changed inside the same `QnResource::update` call. In that case `parentIdChanged` is emitted first (being the base class' signal) and the items are already updated. Then `itemAdded`/`itemRemoved`/`itemChanged` are emitted, and watcher then processed or attempted to process them as duplicate additions or removals.

To fix that I added to the watcher storing of actual item ids (not just resource counts), and ignoring if an existing item is added or if an non-existing item is removed.

Similar problem was with calling `LayoutItemWatcher::removeWatchedLayout` also from a handler that originated from `QnResource::update`.