VMS-51923: Server, PluginManager: Improve error handling when loading plugins
Introduce vms_server_plugin.ini option ignoreLibContextAndSdkVersionFailures (true by default, to
keep the existing behavior). When true, the failure of nxLibContext() or nxSdkVersion() does not
prevent to load the plugin, but the error message is only logged rather than being written into
the PluginInfo (which led to duplicated PluginInfo records).

Additionally, introduce a queryInterface() check for nxpl::PluginInterface on an old-SDK plugin.