VMS-53196: Sync event and message filter settings for the cloud users
For the correct work of the notifications filter on a cross-site layouts we need to sync filter
value between different sites - so the behavior will not differ when user enters another site.

Settings are actually stored on the cloud, desktop client updates them on each change, and then
they are pulled to all sites.