VMS-57401: Fix editing of cloud users
A newly created cloud user had default locale (an empty string). I'm not sure how, as the User Create/Edit dialog does not operate with the "default locale" value, it has actual languages in the combo box and always puts an actual locale string to `UserData` when saving a user.

With an edit attempt, an actual `en_US` was put to `UserData` to replace the default empty string, and it caused a permission check failure on the server.

Since there's no "default locale" concept in the dialog, I kept it that way and implemented a simple check that we're allowed to modify a locale.