VMS-58699: API, Video Download: Deprecate continuousTimestamp parameter - it is always "true" now
The error described in Jira Issue is caused by the code fixing gaps in timestamps in
`ProgressiveDownloadingConsumer::processData()`. Since unconditional fixing gaps in timestamps were
added to `FfmpegMuxer` (see VMS-56892), there is no reason to have the `continuousTimestamps`
parameter in API call and this code in `ProgressiveDownloadingConsumer`. Hence, this parameter is
declared deprecated and the problematic code is removed.