VMS-58314: Fix potential translatable strings errors
Workaround `lupdate` code parsing issues.

Most of the fixes are for the single issue with parsing function return value. If the value type is located in a namespace (e.g. `std::string`), parser could not deduct the translation context. 

I was not able to find the core reason though. Most probably it is caused by some include or code line in a header.

Also added a workaround for the `Private` class, which was declared as a fully functional `QObject` in a separate file.