VMS-58021: Fix Onvif remote archive synchronization when camera has many periods to synchronize.
The problem is that the old algorithm expected to download all recording events within 30 seconds.
However, this may not be possible if a camera has many events. In this case, the synchronization
may not start because the VMS server did not receive complete information about the time periods
available for synchronization. To address this issue, the algorithm of getting recording events
was changed to use the time range filtration parameters of the FindEvents request.