VMS-59407, VMS-59439, VMS-57809: Disable multithreaded rendering with Direct3D and with Metal
Multithreaded rendering with these backends and enabled VSync causes prolonged locks of the GUI tread at the synchronization stage (for 1/60 ms, seems `QQuickWindow`s wait for the main window to swap frames or vice versa). It's either a Qt bug or some interaction between our low-level code and RHI, but I haven't found the reason, and at this point we should just disable multithreaded rendering with those backends.

Also removed an old workaround which is no longer needed (with "We do not rely on Mac OS OpenGL implementation-related throttling" comment).