VMS-59568, VMS-59648: Fix double-paint issue of the Advanced Search dialog
Issue reason is in the conflict of two separate update request mechanism implemented in the Qt
Windows Platform library.

The first mechanism is the update timer, which triggers update every few milliseconds. The second
one is the VSync callback mechanism.

Generally dialog registers in the VSync instantly after creation, so timer is not started. But in
case of “Advanced Search” window the first animation starts before the dialog is registered in the
VSync engine, and QML grid implementation calls forcePolish on each cell height change. But after
dialog is created and registered in VSync, update requests start to come from two different sources
simultaneously, which causes an assert.

Release build has no such issue because this assert is informative only. For us that means in some
cases this dialog can cause high CPU usage when it is in the background.

Workaround delays painting of the most heavy dialog part until the dialog is initialized and ready.