MOBILE-2521: Close socket descriptors in a separate thread on macOS and iOS
On macOS, closing a socket is not instantaneous - e.g. UDP sockets used for multicast may block on close() for 1 second.
So just spawn a separate thread that closes the descriptor to avoid blocking a thread that closes the socket.
Do the same for iOS since we need to terminate within 5 seconds.