VMS-61066 MSVC compiler setting /EHsc added
Related to VMS-61066

The 'Standard C++ exception handling' must be explicitly enabled for Microsoft C++ compiler.

We didn't have it enabled in CMake files, but surprisingly in the most cases it didn't break the build process. But sometimes, for new builds or on new systems or after MSVS update, it fails. So, seems it will be right to add this option now.